Expand STR and NUM to STRING and NUMBER
