+++ packages/php-wasm/universal/src/lib/index.ts
@@ -1,22 +1,16 @@
 export type {
-	DataModule,
-	EmscriptenOptions,
 	FileInfo,
-	RuntimeType,
 	IsomorphicLocalPHP,
 	IsomorphicRemotePHP,
-	PHPLoaderModule,
 	PHPOutput,
 	PHPRunOptions,
-	PHPRuntime,
-	PHPRuntimeId,
 	UniversalPHP,
 	RmDirOptions,
-	WithFilesystem,
-	WithRun,
-	WithRequestHandler,
-	WithPHPIniBindings,
-} from './base-php';
+	HTTPMethod,
+	PHPRequest,
+	PHPRequestHeaders,
+	RequestHandler,
+} from './universal-php';
 
 export { PHPResponse } from './php-response';
 export type { PHPResponseData } from './php-response';
@@ -27,15 +21,22 @@ export {
 	SupportedPHPVersionsList,
 } from './supported-php-versions';
 export type { SupportedPHPVersion } from './supported-php-versions';
-export { BasePHP, loadPHPRuntime, __private__dont__use } from './base-php';
+export { BasePHP, __private__dont__use } from './base-php';
+export { loadPHPRuntime } from './load-php-runtime';
+export type {
+	DataModule,
+	EmscriptenOptions,
+	PHPLoaderModule,
+	PHPRuntime,
+	PHPRuntimeId,
+	RuntimeType,
+} from './load-php-runtime';
 export { rethrowFileSystemError } from './rethrow-file-system-error';
 
-export type {
-	HTTPMethod,
-	PHPRequest,
-	PHPRequestHeaders,
-	PHPRequestHandlerConfiguration,
-} from './php-request-handler';
+export { isLocalPHP } from './is-local-php';
+export { isRemotePHP } from './is-remote-php';
+
+export type { PHPRequestHandlerConfiguration } from './php-request-handler';
 export { PHPRequestHandler } from './php-request-handler';
 export type { PHPBrowserConfiguration } from './php-browser';
 export { PHPBrowser } from './php-browser';
