+++ packages/php-wasm/universal/src/lib/is-local-php.ts
@@ -1,0 +1,8 @@
+import { BasePHP } from './base-php';
+import { IsomorphicLocalPHP, UniversalPHP } from './universal-php';
+
+export function isLocalPHP(
+	playground: UniversalPHP
+): playground is IsomorphicLocalPHP {
+	return !(playground instanceof BasePHP);
+}
