+++ packages/php-wasm/universal/src/lib/is-remote-php.ts
@@ -1,0 +1,8 @@
+import { isLocalPHP } from './is-local-php';
+import { IsomorphicRemotePHP, UniversalPHP } from './universal-php';
+
+export function isRemotePHP(
+	playground: UniversalPHP
+): playground is IsomorphicRemotePHP {
+	return !isLocalPHP(playground);
+}
